Data Seeder

+++ app/Http/Controllers/DestinationController.php
@@ -108,7 +108,9 @@ class DestinationController extends Controller
      */
     public function destroy(Destination $destination)
     {
-        StorageService::delete($destination->image_url);
+        if ($destination->image_url) {
+            StorageService::delete($destination->image_url);
+        }
         $destination->delete();
 
         return response()->json([
